Refactor StatsRecorder
